Extract LargeButton and DoubleButtonCombo components to their own file
