+++ app/src/main/java/com/infomaniak/swisstransfer/ui/components/DoubleButtonCombo.kt
@@ -1,0 +1,76 @@
+/*
+ * Infomaniak SwissTransfer - Android
+ * Copyright (C) 2024 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.infomaniak.swisstransfer.ui.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.infomaniak.swisstransfer.ui.theme.Margin
+
+@Composable
+fun ColumnScope.DoubleButtonCombo(
+    topButton: @Composable (Modifier) -> Unit,
+    bottomButton: @Composable (Modifier) -> Unit
+) {
+    BoxWithConstraints(
+        modifier = Modifier
+            .widthIn(max = 800.dp)
+            .align(Alignment.CenterHorizontally),
+    ) {
+        if (maxWidth < 500.dp) {
+            Column(Modifier.fillMaxWidth()) {
+                topButton(
+                    Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = Margin.Medium)
+                )
+
+                Spacer(modifier = Modifier.height(Margin.Medium))
+
+                bottomButton(
+                    Modifier
+                        .fillMaxWidth()
+                        .padding(start = Margin.Medium, end = Margin.Medium, bottom = Margin.Large)
+                )
+            }
+        } else {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = Margin.Large),
+                horizontalArrangement = Arrangement.spacedBy(Margin.Medium),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                topButton(
+                    Modifier
+                        .weight(1f)
+                        .padding(start = Margin.Medium)
+                )
+
+                bottomButton(
+                    Modifier
+                        .weight(1f)
+                        .padding(end = Margin.Medium)
+                )
+            }
+        }
+    }
+}
